# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Booking Flow API",
    "version": "1.0.0",
    "description": "Local booking demo API (yusnova-flow/backend). Used by automation `swagger:api` via SWAGGER_PATH."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Next.js demo"
    }
  ],
  "paths": {
    "/waste-types": {
      "post": {
        "operationId": "wasteTypes",
        "summary": "List waste type options",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ok"],
                  "properties": {
                    "ok": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/skips": {
      "get": {
        "operationId": "skips",
        "summary": "List skip sizes for a postcode",
        "parameters": [
          {
            "name": "postcode",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "example": "SW1A 1AA" }
          },
          {
            "name": "heavyWaste",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "plasterboard",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "plasterboardOption",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["skips"],
                  "properties": {
                    "skips": {
                      "type": "array",
                      "items": { "type": "object", "additionalProperties": true }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid postcode",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "error": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/postcode/lookup": {
      "post": {
        "operationId": "postcodeLookup",
        "summary": "Look up addresses for a UK postcode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["postcode"],
                "properties": {
                  "postcode": { "type": "string", "example": "SW1A 1AA" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["postcode", "addresses"],
                  "properties": {
                    "postcode": { "type": "string" },
                    "addresses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "line1": { "type": "string" },
                          "city": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing postcode",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "error": { "type": "string" } }
                }
              }
            }
          },
          "500": {
            "description": "Upstream failure (demo)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "error": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/booking/confirm": {
      "post": {
        "operationId": "bookingConfirm",
        "summary": "Confirm a skip booking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["postcode", "skipSize", "price"],
                "properties": {
                  "postcode": { "type": "string", "example": "SW1A 1AA" },
                  "skipSize": { "type": "string", "example": "4-yard" },
                  "price": { "type": "number", "example": 120 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "bookingId"],
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "bookingId": { "type": "string", "example": "BK-12345" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "error": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    }
  }
}